requirements.txt
# Core dependencies for MBABOT
# Using Azure OpenAI Embeddings (no PyTorch required)

# Web Framework
Flask==3.1.2
gunicorn==23.0.0
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
blinker==1.9.0
click==8.3.0

# Database
psycopg2-binary==2.9.10

# Azure OpenAI
openai==1.109.1

# Environment & Config
python-dotenv==1.1.1

# Web Scraping
beautifulsoup4==4.13.5
requests==2.32.5
requests-toolbelt==1.0.0
lxml==6.0.2
soupsieve==2.8
html5lib==1.1
webencodings==0.5.1

# PDF Processing
unstructured==0.18.15
unstructured-client==0.42.3
pdf2image==1.17.0
pdfminer.six==20250506
pypdf==6.1.0
pypdfium2==4.30.0
pikepdf==9.11.0
pillow==11.3.0

# Data Processing
pandas==2.3.2
numpy==2.2.6

# HTTP & Networking
httpcore==1.0.9
httpx==0.28.1
certifi==2025.8.3
urllib3==2.5.0
idna==3.10
charset-normalizer==3.4.3
h11==0.16.0
anyio==4.11.0
sniffio==1.3.1

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
tqdm==4.67.1
colorama==0.4.6
coloredlogs==15.0.1
humanfriendly==10.0

# Text Processing
langdetect==1.0.9
nltk==3.9.1
regex==2025.9.18
RapidFuzz==3.14.1
emoji==2.15.0

# Type Checking & Validation
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.1
pydantic==2.11.9
pydantic_core==2.33.2
annotated-types==0.7.0

# Data Serialization
PyYAML==6.0.2
python-multipart==0.0.20
dataclasses-json==0.6.7
marshmallow==3.26.1

# File Type Detection
filetype==1.2.0
python-magic==0.4.27

# Async I/O
aiofiles==24.1.0

# Security & Crypto
cryptography==46.0.1
cffi==2.0.0
pycparser==2.23

# Compression & Encoding
backoff==2.2.1

# Google Cloud (if needed for OCR)
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.75.0
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
rsa==4.9.1
cachetools==5.5.2

# Misc Dependencies
Deprecated==1.2.18
distro==1.9.0
filelock==3.19.1
fsspec==2025.9.0
jiter==0.11.0
mypy_extensions==1.1.0
packaging==25.0
setuptools==80.9.0
six==1.17.0
wrapt==1.17.3
olefile==0.47
python-iso639==2025.2.18
python-oxmsg==0.0.2
pyreadline3==3.5.4
